--- client/src/components/tournaments/TempSwitch.js
@@ -1,3768 +1,0 @@
-// 8 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "7",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "5",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "1",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 1,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 2,
-                        displayName: "Seed"
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "2",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 3,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 4,
-                        displayName: "Seed"
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "6",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "3",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "4",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 8,
-                        displayName: "Seed"
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 9 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "8",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "6",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "5",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "1",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "2",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 3,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 4,
-                        displayName: "Seed"
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "7",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "3",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "4",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 8,
-                        displayName: "Seed"
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 10 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "9",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "7",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "5",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 1,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "1",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "3",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 8,
-                        displayName: "Seed"
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "8",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "6",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "2",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "4",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 8,
-                        displayName: "Seed"
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 11 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "10",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "8",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "5",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 1,
-                        displayName: "Seed 6"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "1",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed 0"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed 1"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "4",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed 7"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 8,
-                        displayName: "Seed 8"
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "9",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "6",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed 9"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "2",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed 2"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed 3"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "7",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed 10"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "3",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed 4"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed 5"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 12 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "11",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "9",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "5",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 1,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "1",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "6",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "2",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "10",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "7",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "3",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "8",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "4",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 13 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "12",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "10",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "6",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 1,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "1",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "7",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "2",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "3",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "11",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "8",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "4",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "9",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "5",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 14 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "13",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "11",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "7",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 1,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "1",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "8",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "2",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "3",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "12",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "9",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 7,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "4",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "10",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "5",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "6",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 15 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "14",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "12",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "8",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: null,
-                        rank: 1,
-                        displayName: "Seed"
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "1",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "9",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "2",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "3",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "13",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "10",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "4",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "5",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "11",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "6",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "7",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 16 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "15",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "13",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "9",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "1",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "2",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "10",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "3",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "4",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "14",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "11",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "5",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "6",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "12",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "7",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "8",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 17 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "16",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "14",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "10",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "9",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: {
-                                  id: 5,
-                                  scheduled: null,
-                                  name: "1",
-                                  sides: {
-                                    home: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    },
-                                    visitor: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    }
-                                  }
-                                },
-                                rank: 1,
-                                displayName: ""
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "2",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "11",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "3",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "4",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "15",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "12",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "5",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "6",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "13",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "7",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "8",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 18 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "17",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "15",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "11",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "9",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: {
-                                  id: 5,
-                                  scheduled: null,
-                                  name: "1",
-                                  sides: {
-                                    home: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    },
-                                    visitor: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    }
-                                  }
-                                },
-                                rank: 1,
-                                displayName: ""
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "3",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "12",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "4",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "5",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "16",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "13",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "10",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: {
-                                  id: 5,
-                                  scheduled: null,
-                                  name: "2",
-                                  sides: {
-                                    home: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 1,
-                                        displayName: "Seed"
-                                      }
-                                    },
-                                    visitor: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 1,
-                                        displayName: "Seed"
-                                      }
-                                    }
-                                  }
-                                },
-                                rank: 1,
-                                displayName: ""
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "6",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "14",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "7",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "8",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 19 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "18",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "16",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "12",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "9",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: {
-                                  id: 5,
-                                  scheduled: null,
-                                  name: "1",
-                                  sides: {
-                                    home: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    },
-                                    visitor: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    }
-                                  }
-                                },
-                                rank: 1,
-                                displayName: ""
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "4",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "13",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "10",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: {
-                                  id: 5,
-                                  scheduled: null,
-                                  name: "2",
-                                  sides: {
-                                    home: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    },
-                                    visitor: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    }
-                                  }
-                                },
-                                rank: 1,
-                                displayName: ""
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "5",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "17",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "14",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "11",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: {
-                                  id: 5,
-                                  scheduled: null,
-                                  name: "3",
-                                  sides: {
-                                    home: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 1,
-                                        displayName: "Seed"
-                                      }
-                                    },
-                                    visitor: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 1,
-                                        displayName: "Seed"
-                                      }
-                                    }
-                                  }
-                                },
-                                rank: 1,
-                                displayName: ""
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "6",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "15",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "7",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "8",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
-
-
-// 20 Players --------------------------------------------------------
-const game = {
-  id: 1,
-  scheduled: null,
-  name: "19",
-  sides: {
-    home: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "17",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "13",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "9",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: {
-                                  id: 5,
-                                  scheduled: null,
-                                  name: "1",
-                                  sides: {
-                                    home: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    },
-                                    visitor: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    }
-                                  }
-                                },
-                                rank: 1,
-                                displayName: ""
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "5",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "14",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "10",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: {
-                                  id: 5,
-                                  scheduled: null,
-                                  name: "2",
-                                  sides: {
-                                    home: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    },
-                                    visitor: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    }
-                                  }
-                                },
-                                rank: 1,
-                                displayName: ""
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "6",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  }
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    },
-    visitor: {
-      score: { score: 0 },
-      seed: {
-        sourceGame: {
-          id: 2,
-          scheduled: null,
-          name: "18",
-          sides: {
-            home: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "15",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "11",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: {
-                                  id: 5,
-                                  scheduled: null,
-                                  name: "3",
-                                  sides: {
-                                    home: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 1,
-                                        displayName: "Seed"
-                                      }
-                                    },
-                                    visitor: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 1,
-                                        displayName: "Seed"
-                                      }
-                                    }
-                                  }
-                                },
-                                rank: 1,
-                                displayName: ""
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "7",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 1,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            },
-            visitor: {
-              score: { score: 0 },
-              seed: {
-                sourceGame: {
-                  id: 3,
-                  scheduled: null,
-                  name: "16",
-                  sides: {
-                    home: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "12",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: {
-                                  id: 5,
-                                  scheduled: null,
-                                  name: "4",
-                                  sides: {
-                                    home: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    },
-                                    visitor: {
-                                      score: { score: 0 },
-                                      seed: {
-                                        sourceGame: null,
-                                        rank: 2,
-                                        displayName: "Seed"
-                                      }
-                                    }
-                                  }
-                                },
-                                rank: 1,
-                                displayName: ""
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    },
-                    visitor: {
-                      score: { score: 0 },
-                      seed: {
-                        sourceGame: {
-                          id: 4,
-                          scheduled: null,
-                          name: "8",
-                          sides: {
-                            home: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 7,
-                                displayName: "Seed"
-                              }
-                            },
-                            visitor: {
-                              score: { score: 0 },
-                              seed: {
-                                sourceGame: null,
-                                rank: 8,
-                                displayName: "Seed"
-                              }
-                            }
-                          }
-                        },
-                        rank: 1,
-                        displayName: ""
-                      }
-                    }
-                  },
-                },
-                rank: 1,
-                displayName: ""
-              }
-            }
-          }
-        },
-        rank: 1,
-        displayName: ""
-      }
-    }
-  }
-}
